W56-A2 (+A8 fecars part): ALL FOUR PARKED USER ITEMS EXECUTED -- AIHigh_BasicCop dtor NOT-IN-OBJECT -> PASS 8/8 (empty out-of-line base dtor does NOT reroute derived dtors -- catalog receipt); fNumRacers short->int PROVEN (sw/lw word access; size 644 held; screenpost consumer narrowing fix applied); signed-char sweep: fTransitionDirection + fTrackNumber + fCarID x2 structs (17 consumers gated, 2 IMPROVED -4/-6, zero P->F; QImode copy stays lbu either way -- only compare/extend contexts change); $4-CLOBBER LANDED after no-pin-free-route diagnosis (cse tracks VALUE equality, no value-fence forces the register reload; THE SOLE hard-reg-clobber exception, 14-line receipt) -> CalcUsedPrice PASS 67/67; A8: FindSimilarCar 49->39 incl. u_int-cast-poisons-signed-modulo BUG

+++ recon/frontend/common/screenpost.cpp
@@ -197,7 +197,7 @@ void tScreenTournamentStandings::DrawBackground()
   line = 0x2fe;
   tourneyInfo = &tm->fDefinition->fTournaments[
       tm->fDefinition->fTiers[tm->fTier].fTournOffset + tm->fTournament];
-  numRacers = (short)(tm->fNumRacers + (tourneyInfo->fKnockout != 0));
+  numRacers = (short)((short)tm->fNumRacers + (tourneyInfo->fKnockout != 0));
   lastRacer = numRacers - 1;
   for (;;) {
     if (i >= numRacers) {
